+++ processor/src/main/java/arez/processor/ComponentDescriptor.java
@@ -142,7 +142,7 @@ final class ComponentDescriptor
   private final Map<String, ObservedDescriptor> _observeds = new LinkedHashMap<>();
   private final Collection<ObservedDescriptor> _roObserveds =
     Collections.unmodifiableCollection( _observeds.values() );
-  private final Map<ExecutableElement, DependencyDescriptor> _dependencies = new LinkedHashMap<>();
+  private final Map<Element, DependencyDescriptor> _dependencies = new LinkedHashMap<>();
   private final Collection<DependencyDescriptor> _roDependencies =
     Collections.unmodifiableCollection( _dependencies.values() );
   private final Map<VariableElement, CascadeDisposableDescriptor> _cascadeDisposes = new LinkedHashMap<>();
@@ -219,7 +219,8 @@ final class ComponentDescriptor
                                                 isDeprecated( e.getOnDeactivate() ) ||
                                                 isDeprecated( e.getOnStale() ) ) ||
            _observerRefs.values().stream().anyMatch( e -> isDeprecated( e.getMethod() ) ) ||
-           _roDependencies.stream().anyMatch( e -> isDeprecated( e.getMethod() ) ) ||
+           _roDependencies.stream().anyMatch( e -> ( e.isMethodDependency() && isDeprecated( e.getMethod() ) ) ||
+                                                   ( !e.isMethodDependency() && isDeprecated( e.getField() ) ) ) ||
            _roActions.stream().anyMatch( e -> isDeprecated( e.getAction() ) ) ||
            _roObserveds.stream().anyMatch( e -> ( e.hasObserved() && isDeprecated( e.getObserved() ) ) ||
                                                 ( e.hasOnDepsChanged() && isDeprecated( e.getOnDepsChanged() ) ) ) ||
@@ -242,7 +243,7 @@ final class ComponentDescriptor
     return _disposeTrackable;
   }
 
-  private boolean isDeprecated( @Nullable final ExecutableElement element )
+  private boolean isDeprecated( @Nullable final Element element )
   {
     return null != element && null != element.getAnnotation( Deprecated.class );
   }
@@ -1311,6 +1312,22 @@ final class ComponentDescriptor
     ensureNoAbstractMethods( onDepsChangeds.values() );
 
     processCascadeDisposeFields();
+    processComponentDependencyFields();
+  }
+
+  private void processComponentDependencyFields()
+  {
+    ProcessorUtil.getFieldElements( _element )
+      .stream()
+      .filter( f -> null !=
+                    ProcessorUtil.findAnnotationByType( f, Constants.COMPONENT_DEPENDENCY_ANNOTATION_CLASSNAME ) )
+      .forEach( this::processComponentDependencyField );
+  }
+
+  private void processComponentDependencyField( @Nonnull final VariableElement f )
+  {
+    MethodChecks.mustBeSubclassCallable( _element, Constants.COMPONENT_DEPENDENCY_ANNOTATION_CLASSNAME, f );
+    addDependency( f );
   }
 
   private void processCascadeDisposeFields()
@@ -1328,7 +1345,7 @@ final class ComponentDescriptor
     _cascadeDisposes.put( f, new CascadeDisposableDescriptor( f ) );
   }
 
-  private void mustBeCascadeDisposeTypeCompatible( final @Nonnull VariableElement f )
+  private void mustBeCascadeDisposeTypeCompatible( @Nonnull final VariableElement f )
   {
     final TypeElement disposable = _elements.getTypeElement( Constants.DISPOSABLE_CLASSNAME );
     assert null != disposable;
@@ -1402,7 +1419,7 @@ final class ComponentDescriptor
                                       @Nonnull final ObservableDescriptor observable )
   {
     final DependencyDescriptor dependencyDescriptor =
-      _dependencies.computeIfAbsent( method, this::createDependencyDescriptor );
+      _dependencies.computeIfAbsent( method, m -> createMethodDependencyDescriptor( method ) );
     dependencyDescriptor.setObservable( observable );
   }
 
@@ -1930,11 +1947,16 @@ final class ComponentDescriptor
 
   private void addDependency( @Nonnull final ExecutableElement method )
   {
-    _dependencies.put( method, createDependencyDescriptor( method ) );
+    _dependencies.put( method, createMethodDependencyDescriptor( method ) );
+  }
+
+  private void addDependency( @Nonnull final VariableElement field )
+  {
+    _dependencies.put( field, createFieldDependencyDescriptor( field ) );
   }
 
   @Nonnull
-  private DependencyDescriptor createDependencyDescriptor( @Nonnull final ExecutableElement method )
+  private DependencyDescriptor createMethodDependencyDescriptor( @Nonnull final ExecutableElement method )
   {
     MethodChecks.mustNotHaveAnyParameters( Constants.COMPONENT_DEPENDENCY_ANNOTATION_CLASSNAME, method );
     MethodChecks.mustBeSubclassCallable( getElement(), Constants.COMPONENT_DEPENDENCY_ANNOTATION_CLASSNAME, method );
@@ -1963,6 +1985,42 @@ final class ComponentDescriptor
 
     final boolean cascade = isActionCascade( method );
     return new DependencyDescriptor( method, cascade );
+  }
+
+  @Nonnull
+  private DependencyDescriptor createFieldDependencyDescriptor( @Nonnull final VariableElement field )
+  {
+    MethodChecks.mustBeSubclassCallable( getElement(), Constants.COMPONENT_DEPENDENCY_ANNOTATION_CLASSNAME, field );
+    MethodChecks.mustBeFinal( Constants.COMPONENT_DEPENDENCY_ANNOTATION_CLASSNAME, field );
+
+    final TypeMirror type = field.asType();
+    if ( TypeKind.DECLARED != type.getKind() )
+    {
+      throw new ArezProcessorException( "@ComponentDependency target must be a non-primitive value", field );
+    }
+    final TypeElement disposeTrackable = _elements.getTypeElement( Constants.DISPOSE_TRACKABLE_CLASSNAME );
+    assert null != disposeTrackable;
+    if ( !_typeUtils.isAssignable( type, disposeTrackable.asType() ) )
+    {
+      final TypeElement typeElement = (TypeElement) _typeUtils.asElement( type );
+      final AnnotationMirror value =
+        ProcessorUtil.findAnnotationByType( typeElement, Constants.COMPONENT_ANNOTATION_CLASSNAME );
+      if ( null == value || !ProcessorUtil.isDisposableTrackableRequired( _elements, typeElement ) )
+      {
+        throw new ArezProcessorException( "@ComponentDependency target must be an instance compatible with " +
+                                          Constants.DISPOSE_TRACKABLE_CLASSNAME + " or a type annotated " +
+                                          "with @ArezComponent(disposeTrackable=ENABLE)", field );
+      }
+    }
+
+    if ( !isActionCascade( field ) )
+    {
+      throw new ArezProcessorException( "@ComponentDependency target defined an action of 'SET_NULL' but the " +
+                                        "dependency is on a final field and can not be set to null.", field );
+
+    }
+
+    return new DependencyDescriptor( field );
   }
 
   private boolean isActionCascade( @Nonnull final Element method )
@@ -3201,38 +3259,64 @@ final class ComponentDescriptor
       builder.addStatement( "$N.dispose()", GeneratorUtil.DISPOSE_NOTIFIER_FIELD_NAME );
       for ( final DependencyDescriptor dependency : _roDependencies )
       {
-        final ExecutableElement method = dependency.getMethod();
-        final String methodName = method.getSimpleName().toString();
+        final Element element = dependency.getElement();
         final boolean isNonnull =
-          null != ProcessorUtil.findAnnotationByType( method, Constants.NONNULL_ANNOTATION_CLASSNAME );
-        if ( isNonnull )
+          null != ProcessorUtil.findAnnotationByType( element, Constants.NONNULL_ANNOTATION_CLASSNAME );
+
+        if ( dependency.isMethodDependency() )
         {
-          builder.addStatement( "$T.asDisposeTrackable( $N() ).getNotifier().removeOnDisposeListener( this )",
-                                GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-                                methodName );
-        }
-        else
-        {
-          final String varName = GeneratorUtil.VARIABLE_PREFIX + methodName + "_dependency";
-          final boolean abstractObservables = method.getModifiers().contains( Modifier.ABSTRACT );
-          if ( abstractObservables )
+          final ExecutableElement method = dependency.getMethod();
+          final String methodName = method.getSimpleName().toString();
+          if ( isNonnull )
           {
-            builder.addStatement( "final $T $N = this.$N",
-                                  method.getReturnType(),
-                                  varName,
-                                  dependency.getObservable().getDataFieldName() );
+            builder.addStatement( "$T.asDisposeTrackable( $N() ).getNotifier().removeOnDisposeListener( this )",
+                                  GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
+                                  methodName );
           }
           else
           {
-            builder.addStatement( "final $T $N = super.$N()", method.getReturnType(), varName, methodName );
+            final String varName = GeneratorUtil.VARIABLE_PREFIX + methodName + "_dependency";
+            final boolean abstractObservables = method.getModifiers().contains( Modifier.ABSTRACT );
+            if ( abstractObservables )
+            {
+              builder.addStatement( "final $T $N = this.$N",
+                                    method.getReturnType(),
+                                    varName,
+                                    dependency.getObservable().getDataFieldName() );
+            }
+            else
+            {
+              builder.addStatement( "final $T $N = super.$N()", method.getReturnType(), varName, methodName );
+            }
+            final CodeBlock.Builder listenerBlock = CodeBlock.builder();
+            listenerBlock.beginControlFlow( "if ( null != $N )", varName );
+            listenerBlock.addStatement( "$T.asDisposeTrackable( $N ).getNotifier().removeOnDisposeListener( this )",
+                                        GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
+                                        varName );
+            listenerBlock.endControlFlow();
+            builder.addCode( listenerBlock.build() );
           }
-          final CodeBlock.Builder listenerBlock = CodeBlock.builder();
-          listenerBlock.beginControlFlow( "if ( null != $N )", varName );
-          listenerBlock.addStatement( "$T.asDisposeTrackable( $N ).getNotifier().removeOnDisposeListener( this )",
-                                      GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-                                      varName );
-          listenerBlock.endControlFlow();
-          builder.addCode( listenerBlock.build() );
+        }
+        else
+        {
+          final VariableElement field = dependency.getField();
+          final String fieldName = field.getSimpleName().toString();
+          if ( isNonnull )
+          {
+            builder.addStatement( "$T.asDisposeTrackable( this.$N ).getNotifier().removeOnDisposeListener( this )",
+                                  GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
+                                  fieldName );
+          }
+          else
+          {
+            final CodeBlock.Builder listenerBlock = CodeBlock.builder();
+            listenerBlock.beginControlFlow( "if ( null != this.$N )", fieldName );
+            listenerBlock.addStatement( "$T.asDisposeTrackable( this.$N ).getNotifier().removeOnDisposeListener( this )",
+                                        GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
+                                        fieldName );
+            listenerBlock.endControlFlow();
+            builder.addCode( listenerBlock.build() );
+          }
         }
       }
     }
@@ -3584,90 +3668,7 @@ final class ComponentDescriptor
     _roMemoizes.forEach( e -> e.buildInitializer( builder ) );
     _roObserveds.forEach( observed -> observed.buildInitializer( builder ) );
     _roInverses.forEach( e -> e.buildInitializer( builder ) );
-
-    for ( final DependencyDescriptor dep : _roDependencies )
-    {
-      final ExecutableElement method = dep.getMethod();
-      final String methodName = method.getSimpleName().toString();
-      final boolean abstractObservables = method.getModifiers().contains( Modifier.ABSTRACT );
-      final boolean isNonnull =
-        null != ProcessorUtil.findAnnotationByType( method, Constants.NONNULL_ANNOTATION_CLASSNAME );
-      if ( abstractObservables )
-      {
-        if ( isNonnull )
-        {
-          assert dep.shouldCascadeDispose();
-          builder.addStatement( "$T.asDisposeTrackable( $N ).getNotifier().addOnDisposeListener( this, this::dispose )",
-                                GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-                                dep.getObservable().getDataFieldName() );
-        }
-        else
-        {
-          final String varName = GeneratorUtil.VARIABLE_PREFIX + methodName + "_dependency";
-          builder.addStatement( "final $T $N = this.$N",
-                                dep.getMethod().getReturnType(),
-                                varName,
-                                dep.getObservable().getDataFieldName() );
-          final CodeBlock.Builder listenerBlock = CodeBlock.builder();
-          listenerBlock.beginControlFlow( "if ( null != $N )", varName );
-          if ( dep.shouldCascadeDispose() )
-          {
-            listenerBlock.addStatement(
-              "$T.asDisposeTrackable( $N ).getNotifier().addOnDisposeListener( this, this::dispose )",
-              GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-              dep.getObservable().getDataFieldName() );
-          }
-          else
-          {
-            listenerBlock.addStatement(
-              "$T.asDisposeTrackable( $N ).getNotifier().addOnDisposeListener( this, () -> $N( null ) )",
-              GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-              dep.getObservable().getDataFieldName(),
-              dep.getObservable().getSetter().getSimpleName().toString() );
-          }
-          listenerBlock.endControlFlow();
-          builder.addCode( listenerBlock.build() );
-        }
-      }
-      else
-      {
-        if ( isNonnull )
-        {
-          assert dep.shouldCascadeDispose();
-          builder.addStatement(
-            "$T.asDisposeTrackable( super.$N() ).getNotifier().addOnDisposeListener( this, this::dispose )",
-            GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-            method.getSimpleName().toString() );
-        }
-        else
-        {
-          final String varName = GeneratorUtil.VARIABLE_PREFIX + methodName + "_dependency";
-          builder.addStatement( "final $T $N = super.$N()",
-                                dep.getMethod().getReturnType(),
-                                varName,
-                                method.getSimpleName().toString() );
-          final CodeBlock.Builder listenerBlock = CodeBlock.builder();
-          listenerBlock.beginControlFlow( "if ( null != $N )", varName );
-          if ( dep.shouldCascadeDispose() )
-          {
-            listenerBlock.addStatement(
-              "$T.asDisposeTrackable( super.$N() ).getNotifier().addOnDisposeListener( this, this::dispose )",
-              GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-              method.getSimpleName() );
-          }
-          else
-          {
-            listenerBlock.addStatement(
-              "$T.asDisposeTrackable( super.$N() ).getNotifier().addOnDisposeListener( this, () -> $N( null ) )",
-              GeneratorUtil.DISPOSE_TRACKABLE_CLASSNAME,
-              method.getSimpleName(),
-              dep.getObservable().getSetter().getSimpleName().toString() );
-          }
-          listenerBlock.endControlFlow();
-          builder.addCode( listenerBlock.build() );
-        }
-      }
-    }
+    _roDependencies.forEach( e -> e.buildInitializer( builder ) );
 
     GeneratorUtil.setStateForInvariantChecking( builder, "COMPONENT_CONSTRUCTED" );