+++ processor/src/test/resources/bad_input/com/example/component_dependency/SetNullFieldDependency.java
@@ -1,0 +1,12 @@
+package com.example.component_dependency;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentDependency;
+import arez.component.DisposeTrackable;
+
+@ArezComponent( allowEmpty = true )
+public abstract class SetNullFieldDependency
+{
+  @ComponentDependency( action = ComponentDependency.Action.SET_NULL )
+  final DisposeTrackable time = null;
+}