+++ processor/src/test/resources/bad_input/com/example/component_dependency/StaticFieldDependency.java
@@ -1,0 +1,12 @@
+package com.example.component_dependency;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentDependency;
+import arez.component.DisposeTrackable;
+
+@ArezComponent( allowEmpty = true )
+public abstract class StaticFieldDependency
+{
+  @ComponentDependency
+  static final DisposeTrackable time = null;
+}